refactor(insurance): move extract-insurance-card/extract-photo-id out of subscriptions/
They stopped being FHIR subscriptions in an earlier commit on this branch
(now http_auth, called directly by the frontend) but stayed in
src/subscriptions/document-reference/, which no longer describes them.
Moved both, plus their shared extraction-helpers.ts (only ever consumed
by these two), into src/ehr/ alongside rotate-insurance-card-image, the
other EHR-triggered card zambda. process-telemed-recording stays in
subscriptions/document-reference/ — it's still a genuine subscription.

Updated every relative import affected by the depth change, including the
vi.mock() path strings in both test files, rotate-insurance-card-image's
three cross-references, and config/oystehr-core/zambdas.json's src paths.
