+++ config/oystehr-core/zambdas.json
@@ -763,14 +763,14 @@
       "name": "extract-insurance-card",
       "type": "http_auth",
       "runtime": "nodejs22.x",
-      "src": "src/subscriptions/document-reference/extract-insurance-card/index",
+      "src": "src/ehr/extract-insurance-card/index",
       "zip": ".dist/zips/extract-insurance-card.zip"
     },
     "EXTRACT-PHOTO-ID": {
       "name": "extract-photo-id",
       "type": "http_auth",
       "runtime": "nodejs22.x",
-      "src": "src/subscriptions/document-reference/extract-photo-id/index",
+      "src": "src/ehr/extract-photo-id/index",
       "zip": ".dist/zips/extract-photo-id.zip"
     },
     "ROTATE-INSURANCE-CARD-IMAGE": {
