+++ packages/zambdas/src/ehr/card-extraction-shared/extraction-helpers.ts
