+++ packages/zambdas/src/ehr/extract-insurance-card/normalize-image.ts
