+++ packages/zambdas/src/ehr/extract-insurance-card/test/image-fixtures.ts
