+++ packages/zambdas/src/ehr/extract-insurance-card/validateRequestParameters.ts
@@ -1,5 +1,5 @@
 import { DocumentType, ExtractCardInput, INVALID_INPUT_ERROR, MISSING_REQUEST_BODY, Secrets } from 'utils';
-import { ZambdaInput } from '../../../shared';
+import { ZambdaInput } from '../../shared';
 
 // The four card *image* slots. Other titles sharing the 64290-0 type code (e.g. the
 // 'fullInsuranceCard' PDFs) are skipped, not errored — checked against the freshly-fetched
